backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
pdfplumber==0.11.4
python-docx==1.1.2
sentence-transformers==3.0.1
chromadb==0.5.5
groq==0.9.0
pydantic==2.9.2
python-dotenv==1.0.1
pytest==8.3.4
